apps/sylvia_app/requirements.txt
fastapi==0.99.0
uvicorn==0.23.0
pydantic==1.10.9
pyyaml==6.0
matplotlib==3.8.0
plotly==5.16.0
streamlit==1.26.0
